+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Released on 29/08/2022
 - Implemented `/start` command
 - New `/big-release` command
 - New aphorisms
+- Automatizer tasks: new behaviour for latest video/post
+  - From now on, they won't show the LATEST video, but the latest UNSENT OLDEST video/post in the feed.
 
 ## 0.3.2